rework how settings are saved

+++ src/view/PopupApp.vue
@@ -5,7 +5,10 @@
       <li class="nav-item" role="tab">
         <a
           class="nav-link"
-          :class="selectedTab == 'global' ? 'active' : ''"
+          :class="
+            (selectedTab == 'global' ? ' active ' : '') +
+            (loadedTab == 'global' ? '  font-weight-bold ' : '')
+          "
           @click="changeTab('global')"
           >global</a
         >
@@ -13,7 +16,10 @@
       <li class="nav-item" role="tab">
         <a
           class="nav-link"
-          :class="selectedTab == 'domain' ? 'active' : ''"
+          :class="
+            (selectedTab == 'domain' ? ' active ' : '') +
+            (loadedTab == 'domain' ? '  font-weight-bold ' : '')
+          "
           @click="changeTab('domain')"
           >domain</a
         >
@@ -21,7 +27,10 @@
       <li class="nav-item" role="tab">
         <a
           class="nav-link"
-          :class="selectedTab == 'origin' ? 'active' : ''"
+          :class="
+            (selectedTab == 'origin' ? ' active ' : '') +
+            (loadedTab == 'origin' ? '  font-weight-bold ' : '')
+          "
           @click="changeTab('origin')"
           >Origin</a
         >
@@ -30,16 +39,41 @@
         <a
           class="nav-link"
           role="tab"
-          :class="selectedTab == 'page' ? 'active' : ''"
+          :class="
+            (selectedTab == 'page' ? ' active ' : '') +
+            (loadedTab == 'page' ? '  font-weight-bold ' : '')
+          "
           @click="changeTab('page')"
           >Page</a
         >
       </li>
     </ul>
     <!-- Tabs navs -->
-
     <div class="container">
-      <h3>SETTINGS</h3>
+      <div class="row">
+        <div class="col-6"><h3>SETTINGS</h3></div>
+        <div class="col-6">
+          <div class="form-group">
+            <div class="form-group">
+              <div class="btn-group">
+                <button
+                  class="btn-danger btn form-control-sm btn-sm"
+                  @click="deleteSetting"
+                >
+                  Delete settings
+                </button>
+                <button
+                  class="btn-info btn form-control-sm btn-sm"
+                  @click="loadSelectedSettings"
+                >
+                  Refresh
+                </button>
+              </div>
+            </div>
+          </div>
+        </div>
+      </div>
+      <hr class="my-3" />
       <div class="form-group">
         <button
           @click="toggleInjection"
@@ -48,38 +82,39 @@
         >
           {{ enabled ? "disable" : "enable" }}
         </button>
+      </div>
+      <div class="form-group">
         <button
           class="btn-warning btn form-control-sm btn-sm"
           @click="resetDiv"
         >
           Reset overlay position
         </button>
-        <button
-          class="btn-danger btn form-control-sm btn-sm"
-          @click="deleteSetting"
-        >
-          Delete settings
-        </button>
+        {{ selectedSettings?.position }}
       </div>
-      <div class="form-group">
-        <multiselect
-          v-if="pageHeaders"
-          v-model="selectedHeaders"
-          :clear-on-select="false"
-          :options="pageHeaders"
-          :show-labels="false"
-          :multiple="true"
-          tag-placeholder="add"
-          placeholder="Search or add a header"
-          taggable
-          :close-on-select="false"
-          @tag="addTag"
-        >
-        </multiselect>
-        <button :class="'btn btn-outline-success'" @click="saveHeaders">
-          SAVE HEADERS
-        </button>
-      </div>
+    </div>
+    <div class="form-group">
+      <multiselect
+        v-if="selectedSettings.allowedHeaders"
+        v-model="selectedSettings.allowedHeaders"
+        :clear-on-select="false"
+        :options="pageHeaders"
+        :show-labels="false"
+        :multiple="true"
+        tag-placeholder="add"
+        placeholder="Search or add a header"
+        taggable
+        :close-on-select="false"
+        @tag="addTag"
+        @input="() => (this.changed = true)"
+        @remove="() => (this.changed = true)"
+        @select="() => (this.changed = true)"
+      >
+      </multiselect>
+      <button :class="'btn btn-outline-success'" @click="save">SAVE</button>
+    </div>
+    <div class="alert-danger" v-if="changed === true">
+      carefull unsaved changes
     </div>
   </div>
 </template>
@@ -102,19 +137,25 @@ export default {
   name: "popupView",
   data() {
     return {
-      selectedSettings: {},
+      selectedSettings: {
+        inject: false,
+        allowedHeaders: [],
+        position: { x: 0, y: 0 },
+      },
       allowedOrigins: [],
       selectedTab: "origin",
-      activeSettings: {},
+      activeSettings: {
+        inject: false,
+        allowedHeaders: [],
+        position: { x: 0, y: 0 },
+      },
       requestInfo: {},
-      selectedHeaders: [],
+      changed: false,
+      loadedTab: "origin",
     };
   },
   computed: {
     enabled() {
-      console.log("updating selected settings");
-      console.log(this.selectedSettings);
-      console.log(this.selectedSettings?.inject);
       return this.selectedSettings?.inject ?? false;
     },
     pageHeaders() {
@@ -124,14 +165,22 @@ export default {
     },
   },
   methods: {
-    async toggleInjection() {
-      await setSettings(this.selectedTab, !this.enabled);
-      this.pageRefresh();
-      this.loadSelectedSettings();
+    toggleInjection() {
+      this.selectedSettings = {
+        ...this.selectedSettings,
+        inject: !this.enabled,
+      };
+
+      console.log(this.selectedSettings);
+      this.changed = true;
     },
-    async resetDiv() {
-      await setSettings(this.selectedTab, undefined, undefined, { x: 0, y: 0 });
-      this.pageRefresh();
+    resetDiv() {
+      this.selectedSettings.position = { x: 0, y: 0 };
+      this.changed = true;
+      console.log(this.selectedSettings);
+    },
+    changeTab(changeTo) {
+      this.selectedTab = changeTo;
       this.loadSelectedSettings();
     },
     async loadSettings() {
@@ -139,24 +188,34 @@ export default {
         await getMostSpecificSettings();
       this.activeSettings = activeSettings;
       this.selectedTab = selectedTab;
+      this.loadedTab = selectedTab;
       let currentTab = await getCurrentTab();
       this.requestInfo = await getRequestInfo(currentTab.id);
-    },
-    changeTab(changeTo) {
-      this.selectedTab = changeTo;
       this.loadSelectedSettings();
     },
     async loadSelectedSettings() {
-      this.selectedSettings = await getSettings(this.selectedTab);
-      console.log(toRaw(this.selectedSettings));
-      this.selectedHeaders = this.selectedSettings?.allowedHeaders ?? [];
-      console.log(toRaw(this.selectedHeaders));
+      let loadedSettings = await getSettings(this.selectedTab);
+      console.log(loadedSettings);
+      if (!loadedSettings) {
+        this.selectedSettings = {
+          inject: false,
+          allowedHeaders: [],
+          position: { x: 0, y: 0 },
+        };
+        this.changed = true;
+      } else {
+        this.selectedSettings = loadedSettings;
+        this.changed = false;
+      }
+      console.debug(toRaw(this.selectedSettings));
+
+      // console.log(await getAll());
     },
     async deleteSetting() {
+      if (!confirm("Do you really want to delete this settings?")) return;
       await deleteSettings(this.selectedTab);
       console.log("Settings deleted");
 
-      this.loadSelectedSettings();
       this.loadSettings();
     },
     pageRefresh() {
@@ -166,24 +225,23 @@ export default {
       }
     },
     addTag(val) {
-      this.selectedHeaders.push(val);
+      this.selectedSettings?.allowedHeaders.push(val);
     },
-    async saveHeaders() {
+    async save() {
       await setSettings(
         this.selectedTab,
-        undefined,
-        toRaw(this.selectedHeaders)
+        this.selectedSettings.inject,
+        toRaw(this.selectedSettings.allowedHeaders),
+        this.selectedSettings.position
       );
-      if (confirm("Refresh page") === true) {
-        console.log("refreshing page");
-        refreshCurrentPage();
-      }
+      this.pageRefresh();
+      this.loadSelectedSettings();
     },
   },
   mounted() {
-    setTimeout(() => {
-      this.loadSettings();
-      this.loadSelectedSettings();
+    setTimeout(async () => {
+      await this.loadSettings();
+      await this.loadSelectedSettings();
     }, 25);
   },
 };
